+++ config.py
@@ -11,7 +11,7 @@ Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
 Modifié le    : 2026-09-25
-Version       : 0.6.6
+Version       : 0.6.7
 Copyright     : Pierre Théberge
 
 Description
@@ -123,6 +123,7 @@ Modifications
 0.6.4  - 2026-09-24   CR      : Synchronisation de version (aucun changement fonctionnel).
 0.6.5  - 2026-09-25   CR      : Synchronisation de version (aucun changement fonctionnel).
 0.6.6  - 2026-09-25   CR      : Synchronisation de version (aucun changement fonctionnel).
+0.6.7  - 2026-09-25   CR      : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------